requirements.txt
streamlit==0.73.0
tensorflow==2.3.0
numpy==1.18.5
Pillow==7.2.0
